make sensor orientation readable before READY
This, too, matches the behavior of gstahc, a camerabin-compatible
element for Android Camera API. _get_hw() is made to be idempotent, and
the property getting code is made to call this function before giving
the answer.
.
With this patch, QtMultimedia no longers have to set element's state
before reading orientations. However, with the previous patch, it might
still does so, at a cost of a bit slower enumeration.